Refactor main function and modularize codebase
- Refactored the main function for improved readability and maintainability.
- Extracted benchmarking, prediction, and utility functions into separate modules.
- Created dedicated modules for ONNX, OpenVINO, and CUDA operations to enhance code organization.
- Simplified CUDA configurations and added detailed comments for clarity.
